backend/requirements.txt
azure-storage-queue
uuid
